code update

+++ include/tablecache.cpp
@@ -13,11 +13,11 @@ namespace SPA
         return *this;
     }
 
-    int CTable::Find(size_t ordinal, CTable::Operator op, const CComVariant &vt, CTable & tbl) const {
+	int CTable::Find(unsigned int ordinal, CTable::Operator op, const CComVariant &vt, CTable & tbl) const {
         return Find(ordinal, op, (const VARIANT &) vt, tbl);
     }
 
-    int CTable::Find(size_t ordinal, CTable::Operator op, const VARIANT &vt, CTable & tbl) const {
+	int CTable::Find(unsigned int ordinal, CTable::Operator op, const VARIANT &vt, CTable & tbl) const {
         tbl.second.clear();
         tbl.first = first;
         tbl.m_bDataCaseSensitive = m_bDataCaseSensitive;
@@ -25,11 +25,11 @@ namespace SPA
         return -1;
     }
 
-    int CTable::Between(size_t ordinal, const CComVariant &vt0, const CComVariant &vt1, CTable & tbl) const {
+	int CTable::Between(unsigned int ordinal, const CComVariant &vt0, const CComVariant &vt1, CTable & tbl) const {
         return Between(ordinal, (const VARIANT&) vt0, (const VARIANT&) vt1, tbl);
     }
 
-    int CTable::Between(size_t ordinal, const VARIANT &vt0, const VARIANT &vt1, CTable & tbl) const {
+	int CTable::Between(unsigned int ordinal, const VARIANT &vt0, const VARIANT &vt1, CTable & tbl) const {
         tbl.second.clear();
         tbl.first = first;
         tbl.m_bDataCaseSensitive = m_bDataCaseSensitive;
